packages/server-python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-socketio>=5.11.0
httpx>=0.27.0
python-dotenv>=1.0.0
